requirements.txt
streamlit
fastai
pillow
